refactor: move verify_cairo_version_compatibility to utils module
